+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_raih_peduli/screen/view/onBoarding/onboarding.dart';
 import 'package:flutter_raih_peduli/screen/view_model/view_model_onboarding.dart';
 import 'package:provider/provider.dart';
+import 'screen/view/onboarding/onboarding_view.dart';
 import 'screen/view_model/view_model_login.dart';
 
 void main() {
