+++ lib/screen/view_model/view_model_onboarding.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../../model/model_onboarding.dart';
+import '../view/login.dart';
 
 class OnBoardingViewModel with ChangeNotifier {
   int currentIndex = 0;
@@ -16,8 +17,17 @@ class OnBoardingViewModel with ChangeNotifier {
 
   void nextPage(PageController controller) {
     controller.nextPage(
-      duration: const Duration(seconds: 1),
+      duration: const Duration(milliseconds: 500),
       curve: Curves.ease,
+    );
+  }
+
+  void skipPage(BuildContext context) {
+    Navigator.pushReplacement(
+      context,
+      MaterialPageRoute(
+        builder: (_) => const LoginScreen(),
+      ),
     );
   }
 }
